Moving base commands to independent file
